+++ mealie/services/backups/imports.py
@@ -4,9 +4,9 @@ import zipfile
 from pathlib import Path
 from typing import List
 
+from app_config import BACKUP_DIR, IMG_DIR, TEMP_DIR
 from services.recipe_services import Recipe
 from services.settings_services import SiteSettings, SiteTheme
-from app_config import BACKUP_DIR, IMG_DIR, TEMP_DIR
 from utils.logger import logger
 
 
@@ -73,11 +73,6 @@ class ImportDatabase:
             with open(recipe, "r") as f:
                 recipe_dict = json.loads(f.read())
                 recipe_dict = ImportDatabase._recipe_migration(recipe_dict)
-
-                recipe_obj = Recipe(**recipe_dict)
-                recipe_obj.save_to_db()
-                successful_imports.append(recipe.stem)
-                logger.info(f"Imported: {recipe.stem}")
             try:
                 recipe_obj = Recipe(**recipe_dict)
                 recipe_obj.save_to_db()